+++ CHANGELOG.md
@@ -9,9 +9,6 @@
 ### [0.2.5](https://github.com/deanius/antares/releases/tag/v0.2.5)
 * Expose Immutable from Antares
 
-### [0.2.6](https://github.com/deanius/antares/releases/tag/v0.2.6)
-* Send last action, if applicable, once subscribed
-
 ### [0.2.7](https://github.com/deanius/antares/releases/tag/v0.2.7)
 * Underlay parent metadata in return from createConsequence
 * Better logging for store.dispatch errors
@@ -22,3 +19,5 @@
 * Moved .endOfEpic() getter to the return value of 'announce'
 * Antares.originate - creates meta-enhanced action synchronously, same interface as announce
 * Honors the Types config option now - if an action.type has a corresponding validator function, it will be used
+* If subscribe called with a key, client receives current state of that key upon connect (DDP added)
+* Added Antares.startup